+++ src/main/java/com/onz/modules/common/pointHistory/domain/embed/PointInfo.java
@@ -15,7 +15,7 @@ import javax.persistence.Embeddable;
 public class PointInfo {
     @Convert(converter = PointTableConverter.class)
     private PointTable code;
-    private long totAmt;
+    private Integer totAmt;
     private int amt;
     private String description;
